requirements.txt
streamlit
tensorflow
pillow
numpy
google-generativeai
altair<5
